Gapless was disabled in previous patch :/

+++ src/curl/http.cpp
@@ -190,7 +190,7 @@ bool wxCurlHTTP::Post(wxInputStream& buffer, const wxString& szRemoteFile /*= wx
 
         iSize = buffer.GetSize();
 
-        if(iSize == (~(size_t)0))	// wxCurlHTTP does not know how to upload unknown length streams.
+        if(iSize == (int)(~(size_t)0))	// wxCurlHTTP does not know how to upload unknown length streams.
             return false;
 
         SetOpt(CURLOPT_POST, TRUE);
@@ -332,7 +332,7 @@ bool wxCurlHTTP::Put(wxInputStream& buffer, const wxString& szRemoteFile /*= wxE
 
         iSize = buffer.GetSize();
 
-        if(iSize == (~(size_t)0))	// wxCurlHTTP does not know how to upload unknown length streams.
+        if(iSize == (int)(~(size_t)0))	// wxCurlHTTP does not know how to upload unknown length streams.
             return false;
 
         SetOpt(CURLOPT_UPLOAD, TRUE);
